[DF][NFC] Add specialized column reader types
Currently RDF reads all column values via a single RColumnValue type,
which knows how to deal with TTreeReaderValues, TTreeReaderArrays,
datasource columns and defined columns. Not only that's too many hats
for a single type, but this design requires some runtime logical branching
to decide which path to take that can be moved to compile-time.
Finally and most importantly, going forward we want smarter integration
with RDataSource, with as few virtual calls as possible, and we will
require RDataSource-specialized column readers that are templated on the
concrete data source type.
In order to allow this latter use-case and move some decisions to
compile-time, instead of a single RColumnValue type we add specialized
column reader types.

RTreeColumnReader reads values via TTreeReaderValues or TTreeReaderArrays
depending on the class specialization, and RDSColumnReader reads values
from a datasource. A given computation graph will only make use of one
or the other.

RGenericColumnReader provides a virtual interface for cases in which we can only
decide which of the two value readers to use at runtime (namely, in case
users type-erase RDF nodes by casting them to the generic RNode type).

RColumnReaderCRTP decorates the other types with the logic necessary to
deal with Defined columns, which is required in all cases because,
independently on the kind of data source, each column might be a defined
column and we will only find out at runtime.

This commit only adds the new types (hence the [NFC] tag), subsequent
commits will refactor the rest of the code to make use of them.