[DF] Template RLoopManager over the kind of data source
I.e. rename RLoopManager to RLoopManagerBase and add a templated
derived class. All logic remains, unchanged, in RLoopManagerBase.

We need RLoopManager to carry that extra template parameter so all
nodes can tell, at compile time, whether we are reading from a TTree
or a concrete RDataSource implementation, and in the latter case which
is the concrete data-source type.
This is in preparation of column readers that can be specialized
on the data source type, to avoid virtual calls.

By casting a dataframe node to the generic RNode type, virtualization
is re-introduced and all downstream nodes in the graph lose the
information on the concrete data-source type.