[DF] Propagate DataSource type through the computation graph
So that each node of the computation graph can build column readers
with knowledge of the exact type of data source (TTree or a concrete
RDataSource implementation). In presence of virtualization, e.g.
due to jitting or an explicit cast to RNode on the part of a user,
DataSource_t is `void`, which indicates the information is not known
at compile-time.